requirements.txt
Django~=5.1.7
django-rest-framework=3.15.2
django-unfold~=0.52.0
dotenv~=0.9.9
psycopg2=2.9.10
python-dotenv~=1.0.1